feat: wagmi logic in hook + nonce
